# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fleet visibility, session history, and control plane for AI agent fleets",
        "title": "Flightdeck Query API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/v1/policies": {
            "get": {
                "description": "Returns all token policies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policies"
                ],
                "summary": "List policies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.Policy"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new token policy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policies"
                ],
                "summary": "Create policy",
                "parameters": [
                    {
                        "description": "Policy definition",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.Policy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/policies/{id}": {
            "put": {
                "description": "Updates an existing token policy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policies"
                ],
                "summary": "Update policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Policy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Policy definition",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Policy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a token policy",
                "tags": [
                    "policies"
                ],
                "summary": "Delete policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Policy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/policy": {
            "get": {
                "description": "Returns the most specific policy for a given flavor/session scope. Lookup order: session \u003e flavor \u003e org",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policies"
                ],
                "summary": "Get effective policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent flavor",
                        "name": "flavor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Policy"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.PolicyRequest": {
            "type": "object",
            "properties": {
                "block_at_pct": {
                    "type": "integer"
                },
                "degrade_at_pct": {
                    "type": "integer"
                },
                "degrade_to": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "scope_value": {
                    "type": "string"
                },
                "token_limit": {
                    "type": "integer"
                },
                "warn_at_pct": {
                    "type": "integer"
                }
            }
        },
...
        "store.Policy": {
            "type": "object",
            "properties": {
                "block_at_pct": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "degrade_at_pct": {
                    "type": "integer"
                },
                "degrade_to": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "scope_value": {
                    "type": "string"
                },
                "token_limit": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "warn_at_pct": {
                    "type": "integer"
                }
            }
        },
...
    }
}